Fix x86_64 (LLP64) build: void* attribute slot uses intptr_t, not long
ItemAttributes stores a uint32_t in a void* slot; getIntAttr/increaseIntAttr cast
it via (long), which loses precision on Windows x86_64 (LLP64: long is 32-bit,
pointers 64-bit). intptr_t is pointer-sized everywhere; the stored values are 32-bit
so behaviour is identical (LP64 long==pointer hid this on Linux/macOS).

+++ source/item.cpp
@@ -890,7 +890,7 @@ uint32_t ItemAttributes::getIntAttr(itemAttrTypes type) const
 		
 	Attribute* attr = getAttr(type);
 	if(attr){
-		return (uint32_t)(long)attr->value;
+		return (uint32_t)(intptr_t)attr->value;
 	}
 	else{
 		return 0;
@@ -915,7 +915,7 @@ void ItemAttributes::increaseIntAttr(itemAttrTypes type, uint32_t value)
 	
 	Attribute* attr = getAttr(type);
 	if(attr){
-		attr->value = (void*)((long)attr->value + value);
+		attr->value = (void*)((intptr_t)attr->value + value);
 	}
 }
